backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # For PostgreSQL (optional)
# timescaledb==2.0.0  # For TimescaleDB (optional)

# Redis & Task Queue
redis==5.0.1
celery==5.3.4

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
playwright==1.40.0
requests==2.31.0
selenium==4.15.2

# ML & Data Science
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.3
numpy==1.26.2

# NLP & Sentiment Analysis
transformers==4.35.2
torch==2.1.1
textblob==0.17.1
vaderSentiment==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
pydantic-extra-types==2.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

